PepperMount overhaul.
This change fixes a couple of bugs in PepperMount:
- It is no longer possible to create a pp::FileIO on a non-main thread.
- FileIO::Query never returns PP_FILETYPE_DIRECTORY because you can only get a FileIO for regular files.

And implements new functionality:
- Mkdir
- Getdents via a pluggable DirectoryReader implementation.

The problem with Getdents is the lack of stable interface for listing directories in Pepper.
Therefore, we let the user provide their own implementation of that.
One way to do that is by sending a message to the JS side.
Review URL: http://codereview.chromium.org/8503011