Simplify save/load experience so that user won't need to remember saving (or loading) to localStorage.
Thanks to PostMessage for supporting sending ArrayBuffer, we don't need to encode the blob now.  With that, I made the "save game" command to write through to localStorage.  Previously user needs to 1) tell SNES to save the game, then 2) press "save" button on the page to copy to localStorage.

There are some (hacky?) protocol to copy from localStorage to SNES (and vice versa).  For example, when the user presses "shift-1" to save the game, javascript postMessage's to main NaCl thread, which insert the event to event queue.  The new change here is to let main thread sending an ACK message to javascript, which will consequently tell main NaCl thread to send the save file back to store in localStorage.

For loading a save, it's basically not change.  But when loading from localStorage for the first time (notmally after start SNES), javascript upload the blob to main NaCl thread first, and then send "load snapshot" command to SNES.
Review URL: https://chromiumcodereview.appspot.com/10237001