Altering MainThreadRunner to be more performant.
MainThreadRunner currently issues a RunOnMainThread
ppapi call for each request.
This is to ensure the main thread handles the request in a
timely fashion.
Instead, we should only send a wakeup in the event that
the run queue is empty (as there is no pending wakeup).
Then on the main thread we should handle all pending
requests as they are intended to be small non-blocking operations.

BUG=None
TEST=None
R=eugenis@google.com,jeffbailey@google.com
Review URL: https://chromiumcodereview.appspot.com/10917030