Use -isystem for naclports include.
GCC is a little more forgiving when generating warnings for
system includes.  Using -isystem for naclports means that
projects can build with -Wall without getting warnings from
the naclports headers.

The particular case warning that triggered me to write
this change was a bunch of missing enum in switch warnings
in the vorbis headers.

BUG=
R=binji@chromium.org

Review URL: https://codereview.chromium.org/14735004