Fix pnacl build of webp.
Be explicit about webp depending on jpeg and tiff and
add -lm which they in turn depend on.

Not quite sure how this was broken, my guess is that
libtool .la magic was effected by the recent change to
install pnacl libs and headers directly into the toolchain.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/19428002