Remove -m32 from host builds for nethack.
Fear the bitage differences would matter for host tools with nethack lead me
to compile host binaries for it with gcc -m32. Later experience has lead me
to believe this is no longer needed.

Nethack for newlib was recently enabled on the buidlers.
Shuffling of the builds in the meantime has resulted in nethack being assigned
to a builder without multilib support. Thus causing the gcc -m32 build to fail.

BUG=None
TEST=local
R=binji@chromium.org

Review URL: https://codereview.chromium.org/20785002