Build and run the openssl unittests.
Background: An external developer is seeing slowdown in
the NaCl builds compared to the host builds. One hypothesis
is that NaCl's ILP32 model means that even on x86-64
machines the NaCl build does not do 64-bit loads-stores
so we see a hit with that. Before doing anything drastic
with the build, start running some unit tests!
(preliminary tests show that 64-bit arith can cut the
running time of a benchmark by 30%).

For Glibc, we just need to set up the sel_ldr scripts.
Ideally this would build the tests, write the sel_ldr
scripts, then run "make test" from the top level.  However,
writing the sel_ldr scripts require we know the list of
tests still... unless the compiler were to generate scripts
as it builds.

For Newlib, we need to configure the build differently so
that "-ldl" is not used to link (won't work that way!).
Some of the tests won't link without socket() and
shutdown(), so skip those for now. A couple of them also
segfault with both nacl-gcc and pnacl.

BUG= https://code.google.com/p/naclports/issues/detail?id=88
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/37463005