Remove use of libnosys.
libnosys.a has been deprecated as of SDK version 31.  These
days libc provides all the NOSYS stubs that libnosys used
to provide.  While the library still exists as an empty
archive we can simply a lot of the naclports by removing
all referecnes to it.

This change means that many ports now require pepper_31
so I've set this as the default MIN_SDK_VERSION.  It should
still be possible for build some ports with older SDKs
but this is now generally supported.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/86433002