Minor cleanup of naclports build system.
This change removes some of the boilerplate from
the individual build.sh file for the ports.
Specifically it removes calls to Banner which are
now handled in common.sh.  It also removes as many
of the custom PackageInstall functions as possible.
There are three remaining that still need to be
removed somehow.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/139743015