Fixing gdb slow start and dropping nethack specific code.
Gdb uses libiberty which makes the assumption that vfork calls actually work.
It compounds this mistake by using an invocation pattern for vfork in which it
sleeps and retries in the event of failure. As vfork does not currently work,
disabling the code path for a faster start.

BUG=None
TEST=forthcoming (other CL)
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/208423009