Rename ports that include underscore in their name.
This change is in preparation for a change that outlaws
underscore in package names and versions so that they can
be used a special character in the file name of binary
packages (just like debian/ubuntu packages).

Also, rename webp to libwebp to the directory name matches
the package name.

This change is mostly pure rename operations.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/213643002