Use manifest 'key' to connect debugger extension to GDB app.
Now that we've pushed our gdb app to the web store, we can use the key option
in manifest.json to allow the app to use its real extension id even for
testing.

Dropping use of the test extension for this purpose.

Fix typo in comment.

R=binji@chromium.org
BUG=None
TEST=gdb_test + debug_test

Review URL: https://codereview.chromium.org/263263002