When installing packages replace existing version if it is not identical.
We currently write the contents of pkg_info to the stamp
folder, but we were not checking it, so requesting install
would do nothing, as long as some version was already
installed.

In addition add CONFIG=debug/release to the end of the stamp
file so that debug and release versions of a package replace
each other.

This should fix the SDK bundle builder.

R=binji@chromium.org, binji

Review URL: https://codereview.chromium.org/309573003