Fix pandas static linking.
The pandas build had a number of minor issues which this
corrects.  The namespace for static libraries was getting
confused with the namespace for relative imports, which this
fixes.  msgpack, which is disabled for network interface
reasons, was getting imported and breaking global import.
And finally, there were name collisions for both tokenizer.o
and parser.o, which this fixes.

R=sbc@chromium.org

Review URL: https://codereview.chromium.org/299413002