Make bash, thus devenv build with newlib and pnacl.
Fixed bash build to work with pnacl and newlib.
Disabled attempt to set file time in newlib build of unzip (so it can be
actually used).

Repartitioned buildbots.

Build and runs with tweaks to startup, but has various issues.
These will be addressed in a forthcoming refactor of the devenv bashrc.

BUG=None
TEST=local
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/335143007