Fix ruby compile-time prefix.
This fixes a recent regression causes by a change the
default prefix used to configure packages.

Ruby bakes the prefix in at compile time so use the real
prefix (in this case /) when configuring it.  Without this
our dummy prefix is baked in and of course is not correct at
runtime.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/349183003