Add getpgid()
Add getpgid(). This does not cover the full range of possible errors yet,
since we are not yet tracking sessions or parent-child relationships (which
we should really add for waitpid(), though then we have to decide what to do
when waitpid is directly called from javascript).

BUG=None
R=binji@chromium.org, bradnelson@google.com

Review URL: https://codereview.chromium.org/458663002