Strip binaries in PublishByArchForDevEnv.
I noticed that the coreutils zip file was pretty huge
and that it contained unstripped binaries.

The change causes PublishByArchForDevEnv to strip
all binaries before zipping them zip.

An alternative would be to use the 'install-strip'
target that automake generates, but that would be
a larger change and would only work with automake
packages.

Also removed the custom build step from coreutils
as it appears to be doing nothing (arguments to
DefaultBuildStep are ignored).

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/504123003