Disabling broken ports to green the tree.
xkeyboard-config is currently a runtime only dependency of xorg-server.
It doesn't build on the bots for lack of a fresher intltool.
Disabling everywhere.

Disabling xkbcomp on the mac bots because it doesn't build there.
Will investigate.

Disabling xfonts + the fonts on mac because the lack of mkfontdir there.

BUG=None
TEST=None
TBR=sbc@chromium.org

Review URL: https://codereview.chromium.org/541343002